backend/requirements.txt
fastapi
uvicorn
tensorflow
scikit-learn
joblib
Pillow

fastapi
uvicorn
joblib
tensorflow
numpy
scikit-learn
pillow
python-multipart
